### 2024-11-21 02d8de7
docs: add doi to paper.bib
