+++ paper.bib
@@ -5,7 +5,8 @@
   volume={24},
   pages={935--942},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10822-010-9385-x}
 }
 
 @article{ropp2019dimorphite,
@@ -15,7 +16,8 @@
   volume={11},
   pages={1--8},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13321-019-0336-9}
 }
 
 @article{reis2022fast,
@@ -26,7 +28,8 @@
   number={8},
   pages={5068--5078},
   year={2022},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.jctc.2c00308}
 }
 
 @article{cock2009biopython,
@@ -37,16 +40,18 @@
   number={11},
   pages={1422},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btp163}
 }
 
 @misc{rdkit,
-  author = {Landrum, Greg},
-  title = {RDKit},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/rdkit/rdkit}
+  author={Landrum, Greg},
+  title={RDKit},
+  year={2024},
+  publisher={GitHub},
+  journal={GitHub repository},
+  url={https://github.com/rdkit/rdkit},
+  doi={10.5281/zenodo.591637}
 }
 
 @article{o2011open,
@@ -56,7 +61,8 @@
   volume={3},
   pages={1--14},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/1758-2946-3-33}
 }
 
 @misc{meeko,
