### 2024-11-21 bb0b3dc
docs: revert RDKit DOI
