+++ paper.bib
@@ -50,8 +50,7 @@
   year={2024},
   publisher={GitHub},
   journal={GitHub repository},
-  url={https://github.com/rdkit/rdkit},
-  doi={10.5281/zenodo.591637}
+  url={https://github.com/rdkit/rdkit}
 }
 
 @article{o2011open,
