### 2025-01-16 #7511 @radifar
Thank you, we appreciate that. As for, the license, we already replace it with the full standard text as requested.
