### 2025-01-18 c064992
style: add ruff
