+++ tests/conftest.py
@@ -6,234 +6,256 @@ from rdkit import Chem
 
 pytest.register_assert_rewrite("helpers")
 
+
 @pytest.fixture
 def mock_config_default():
     config_default = Mock()
     config_default.load.return_value = "configuration"
     config_default.data = dict(
-        input = "tests/data/6hsh.pdb",
-        protein_chain = "A",
-        ligand_chain = "A",
-        ligand_id = "GOK",
-        protein_out = "protein_6hsh.pdb",
-        ligand_out = "GOK.pdb",
-        ligand_image = "GOK.png",
+        input="tests/data/6hsh.pdb",
+        protein_chain="A",
+        ligand_chain="A",
+        ligand_id="GOK",
+        protein_out="protein_6hsh.pdb",
+        ligand_out="GOK.pdb",
+        ligand_image="GOK.png",
     )
 
     return config_default
+
 
 @pytest.fixture
 def mock_config():
     config = Mock()
     config.load.return_value = "configuration"
     config.data = dict(
-        input = "tests/data/6hsh.pdb",
-        protein_chain = "A",
-        ligand_chain = "A",
-        ligand_id = "GOK",
-        include_metal = "yes",
-        protein_out = "protein_6hsh.pdb",
-        ligand_out = "GOK.pdb",
-        ligand_image = "GOK.png",
-        image_size = "large",
-        ligand_smiles = "Cn1cc(c2c1cccc2)CNCC3CCN(CC3)c4ncc(cn4)C(=O)NO",
-        ph = 7.4
+        input="tests/data/6hsh.pdb",
+        protein_chain="A",
+        ligand_chain="A",
+        ligand_id="GOK",
+        include_metal="yes",
+        protein_out="protein_6hsh.pdb",
+        ligand_out="GOK.pdb",
+        ligand_image="GOK.png",
+        image_size="large",
+        ligand_smiles="Cn1cc(c2c1cccc2)CNCC3CCN(CC3)c4ncc(cn4)C(=O)NO",
+        ph=7.4,
     )
 
     return config
+
 
 @pytest.fixture
 def mock_config_multichain_protein():
     config = Mock()
     config.load.return_value = "configuration"
     config.data = dict(
-        input = "tests/data/6hsh.pdb",
-        protein_chain = "A B",
-        ligand_chain = "A",
-        ligand_id = "GOK",
-        include_metal = "yes",
-        protein_out = "protein_6hsh.pdb",
-        ligand_out = "GOK.pdb",
-        ligand_image = "GOK.png",
-        image_size = "large",
-        ligand_smiles = "Cn1cc(c2c1cccc2)CNCC3CCN(CC3)c4ncc(cn4)C(=O)NO",
-        ph = 7.4
+        input="tests/data/6hsh.pdb",
+        protein_chain="A B",
+        ligand_chain="A",
+        ligand_id="GOK",
+        include_metal="yes",
+        protein_out="protein_6hsh.pdb",
+        ligand_out="GOK.pdb",
+        ligand_image="GOK.png",
+        image_size="large",
+        ligand_smiles="Cn1cc(c2c1cccc2)CNCC3CCN(CC3)c4ncc(cn4)C(=O)NO",
+        ph=7.4,
     )
 
     return config
+
 
 @pytest.fixture
 def mock_config_multiligand():
     config = Mock()
     config.load.return_value = "configuration"
     config.data = dict(
-        input = "tests/data/3ccw.pdb",
-        protein_chain = "A",
-        ligand_chain = "B",
[…]
