+++ src/pilah/pKAI/protein.py
@@ -15,7 +15,6 @@ PK_MODS = {
 
 class Protein:
     def __init__(self, pdb_f):
-
         self.pdb_f = pdb_f
         self.tit_residues = {}
         self.all_residues = {}
@@ -84,7 +83,7 @@ class Protein:
 
                     if resnumb not in self.tit_residues[chain]:
                         self.tit_residues[chain][resnumb] = new_res
-    
+
     def read_pdb(self):
         """Removes all atoms from pdb_f that are not Nitrogens, Sulfurs, or Oxygens"""
         tit_aas = PK_MODS.keys()
@@ -144,7 +143,6 @@ class Protein:
         return (aname, anumb, b, resname, chain, resnumb, x, y, z, icode)
 
     def apply_cutoff(self, cutoff_dist=15):
-
         # TODO speed up by doing distance matrix a priori in cython
 
         for residue in self.iter_residues(titrable_only=True):
