### 2025-01-18 c9d256f
style: complete ruff configuration
