### 2025-01-23 #7511 @Kevin-Mattheus-Moerman
@fboehm can you handle this one?
