### 2025-01-23 #7511 @fboehm
@moritzobenauer @patrickfuchs @gpantel @mortele @dominiquesydow - would you be willing to review this submission for JOSS? My name is Fred, and I'm the editor for this submission. It seems to overlap with your expertise. Please let me know if you have any questions.
