### 2025-02-01 #7720 @moritzobenauer
Dear authors,

Thank you for submitting this repository to JOSS. The following comments are categorized as **Minor Revisions** needed and **Major Revisions** needed following the JOSS guidelines.

### Major Revisions:

**Quality of writing.** While we all acknowledge that it is difficult to write a proper journal article (especially so if it is not your native language) this manuscript suffers from major grammatical and ellipsis errors. This makes the manuscript hard to read and understand, and, therefore, detrimental to reaching a larger audience. The software has scientific merit and it would be a shame if the quality of the manuscript prevented new users from using it.

**Installation instructions:** You describe the installation of the binary file. Please add a step-by-step installation guide to set up the correct virtual environment and download the repo so that you can run the software without downloading the binary executable. This would allow Windows and Mac users to run the software and `pilah` would reach a broader audience.

**Automated tests:** Please include detailed instructions for running the included Python tests along with the installation instructions mentioned above. Even a novice user without experience should be able to follow these steps and be able to verify the functionality of the software. According to the JOSS guidelines, a Github actions integration of these tests would be perfect, but is not necessary, however.

**A statement of need:** Please add a statement of need to your readme file. You can use a shortened version of the one from your manuscript.

### Minor Revisions:

**Contribution and authorship:** According to the repo's commit history only M. Radifar committed to this repo. What was the role of the article's second author E. P. Istyastono? Please not that according to JOSS policy, purely financial contributions are not considered sufficient for co-authorships.

**Comments on the Manuscript:**

1) It is best practice to cite all of the software used and mentioned in the main text of the manuscript instead of the acknowledgments. Also, Figure 1 needs a proper caption describing the workflow.

2) Is it planned to include a command line interface to `pilah` so that users can input options via flags instead of a text file? Looking at your parser structure this should be straight forward and would allow for high-throughput analysis in the future.

3) You mention that competing software suffers from bad accuracy of the bond order assignment. How does `pilah` compare to _SPORES_ regarding this?

4) Please elaborate further on the deleting of residues with missing atoms or wrong bond lengths. Is this common practice in competing software? I would suspect that deleting individual residues might alter the resulting protein significantly, hence influencing the chemical/physical properties.

### A few minor things/tips:

Since your docs are more detailed than your readme file you could link the docs a bit more prominently in your readme file.

If you want to make the software more accessible to people with no coding background, you should try to catch common parser errors (e.g. entering a string instead of a float for pH, wrong SMILES code, etc.) instead of outputting the _Python_ error messages. This might be helpful to people with no substantial background knowledge.

Thank you for contributing this software to the community.

Best,
Moritz
