### 2025-02-15 radifar/pilah#37 @dominiquesydow
Help with local development setup
Hello @radifar,

I am one of the reviewers of your JOSS submission over at https://github.com/openjournals/joss-reviews/issues/7720.

When I am following your instructions here:
https://pilah.readthedocs.io/en/latest/contributing.html#get-started

I am not familiar with `poetry` and have some problems getting started, can you please help? (macOS, M1)

1. Downloaded the repo and cd into `/path/to/pilah`
2. `poetry env use 3.11`
3. `poetry install`
4. `poetry run pilah`

The last command results in
```
Traceback (most recent call last):
  File "/Users/dominique/Library/Caches/pypoetry/virtualenvs/pilah-_jBZ_YJB-py3.11/bin/pilah", line 3, in <module>
    from pilah.pilah import app
ModuleNotFoundError: No module named 'pilah.pilah'
```

As I said, might just be me missing something on how things are done with `poetry`, thanks!
