### 2025-02-15 #7720 @dominiquesydow
Dear authors,

Thanks a lot for submitting your work! I can see that PiLAH is a helpful workflow tool that brings together a lot of relevant tasks for protein-ligand preparation.

Unfortunately, I am having trouble running the software. I am on a Mac, so cannot run the executable, tried the local dev setup but (possibly due to unfamiliarity with `poetry`) failed so far. Opened issue to get help at https://github.com/radifar/pilah/issues/37 and _will test software for this review once my setup was successful_.

@moritzobenauer, you could confirm that the executable is running fine? A quick check on my old linux was unsuccessful but that may just be down to an too old Ubuntu version (which I am not going to fix right now).

---

As per [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#what-happens-if-the-software-i-m-reviewing-doesn-t-meet-the-joss-criteria), I am also going to group my comments into major and minor and you will find that I am agreeing with all the points that @moritzobenauer has raised already but I will repeat them nonetheless.

### Major Revisions

#### Installation + Tests

- Please add detailed end-to-end installation steps to run the tool without the executable. As @moritzobenauer said, this opens up not only dev opportunities but also allows MacOS (like myself) and Windows users to interact with the tool. Currently the installation is a bit hidden [here](https://pilah.readthedocs.io/en/latest/contributing.html#get-started), I would move it actually to the main installation section (or redirect MacOS and Windows users here).
- Please add documentation on how to run the tests (ideally you would hook those up with the CI 🙏 - currently the CI only contains a PDF autogenerator for the manuscript).

#### Manuscript

- Language: Please rework the whole manuscript (plus README and documentation) grammatically, incl sometimes missing verbs in a sentence, thank you!
- Cite all the software packages and models mentioned in the manuscript (those you actually used but also those that you only mention).
- Figure 1: Appreciate the graphic workflow overview! Please add a detailed caption (minor: update "Fix insertion" with something more meaningful)

### Minor Revisions

- Can you please confirm that Enade Perdana Istyastono meets [JOSS' co-authorship requirements](https://joss.readthedocs.io/en/latest/submitting.html#authorship)? Muhammad Radifar is the only repo contributor visible to us via the commit history.
- As per JOSS guidelines, the Statement of Need needs to be added to the documentation.
- Towards the end of this review process, I suggest to cut a new release that goes beyond a beta release (currently: 0.1.0b2)
- State of the Field: You mention SPORES as a closed-sourced tool from 2010. Is there other relevant software beyond SPORES that needs mentioning?
- I suggest to not duplicate installation instructions in
[README](https://github.com/radifar/pilah?tab=readme-ov-file#installation) and
[documentation](https://pilah.readthedocs.io/en/latest/usage/installation.html)
but link in README back to documentation; duplication between README and docs can be a real pain to maintain - thanks!

Many thanks for open-sourcing your work!

Kind regards,
Dominique
