### 2025-02-17 #7720 @moritzobenauer
Hey @dominiquesydow ,
I tested the software on Ubuntu 24.04 without any issues. All the functionalities work as reported!
