### 2025-02-20 #7720 @radifar
Dear reviewers and editor,

My apology for the late update from me. I've got health problem since early January and just got recovered recently.

First, I would like to address @dominiquesydow comment on PiLAH usage on Mac. Actually in the future we planned to package PiLAH using [fpm](https://fpm.readthedocs.io/en/latest/getting-started.html) so that it can be installed in Linux (as deb or rpm package), Windows, and Mac (pkg package). In the mean time, to quickly address this issue I would add the instruction on how to setup the development environment on Mac and Linux.

As for the major and minor revisions, many thanks for your thoughtful input. I have begun discussing these with my co-author and will give our response ASAP.

Best,
Radif
