### 2025-02-20 radifar/pilah#37 @radifar
Hi Dominique,

My bad, the truth is the local development environment setup is generated with the [Python Packages](https://github.com/py-pkgs/py-pkgs) cookiecutter which is using Poetry as its package and dependency manager. But due to recent progress in the Python packaging practice, I decided to switch to [uv](https://docs.astral.sh/uv/) as the package and dependency manager on January. And I haven't updated the contributing page.

I would be working on the local development environment setup soon, hopefully you can try and test it on your Mac by this weekend.

Best,
Radif
