### 2025-07-11 #7720 @editorialbot
Paper rejected.
